Rename session/common/netutils package to netutils
