+++ session/common/netutils/errors.go
@@ -16,7 +16,7 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-package utils
+package netutils
 
 import (
 	"context"
